Fixing bug in hex serialization.
