+++ client-py/casperlabs_client/commands/balance_cmd.py
@@ -6,10 +6,7 @@ from casperlabs_client.decorators import guarded_command
 NAME: str = "balance"
 HELP: str = "Returns the balance of the account at the specified block."
 OPTIONS = [
-    [
-        ("-a", "--address"),
-        dict(required=True, type=str, help="Account's public key in hex."),
-    ],
+    [("-a", "--address"), dict(required=True, type=str, help="Account's hash in hex.")],
     [
         ("-b", "--block-hash"),
         dict(required=True, type=str, help="Hash of the block to query the state of"),
