+++ client-py/casperlabs_client/commands/query_state_cmd.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from casperlabs_client import CasperLabsClient, utils
+from casperlabs_client import CasperLabsClient, reformat
 from casperlabs_client.decorators import guarded_command
 
 
@@ -43,4 +43,4 @@ def method(casperlabs_client: CasperLabsClient, args: Dict):
     response = casperlabs_client.query_state(
         args.get("block_hash"), args.get("key"), args.get("path", ""), args.get("type")
     )
-    print(utils.hexify(response))
+    print(reformat.hexify(response))
