+++ client-py/casperlabs_client/consts.py
@@ -5,7 +5,10 @@ STATUS_CHECK_DELAY = 0.5
 STATUS_TIMEOUT = 180  # 3 minutes
 DEFAULT_PAYMENT_AMOUNT = 10000000
 VISUALIZE_DAG_STREAM_DELAY = 5
-SUPPORTED_KEY_ALGORITHMS = ("ed25519", "bn2254")
+
+ED25519_KEY_ALGORITHM = "ed25519"
+BN2254_KEY_ALGORITHM = "bn2254"
+SUPPORTED_KEY_ALGORITHMS = (ED25519_KEY_ALGORITHM, BN2254_KEY_ALGORITHM)
 
 BUNDLED_TRANSFER_WASM: str = "transfer_to_account_u512.wasm"
